move to compat.immutable
