+++ src/main/scala-2.11_2.12/scala/collection/compat/CompatImpl.scala
@@ -1,8 +1,8 @@
-package scala.collection
-package compat
+package scala.collection.compat
 
 import scala.collection.generic.CanBuildFrom
 import scala.collection.mutable.Builder
+import scala.collection.{immutable => i, mutable => m}
 
 private[compat] object CompatImpl {
   def simpleCBF[A, C](f: => Builder[A, C]): CanBuildFrom[Any, A, C] = new CanBuildFrom[Any, A, C] {
@@ -10,6 +10,6 @@ private[compat] object CompatImpl {
     def apply(): Builder[A, C] = f
   }
 
-  type ImmutableBitSetCC[X] = ({ type L[_] = immutable.BitSet })#L[X]
-  type MutableBitSetCC[X] = ({ type L[_] = mutable.BitSet })#L[X]
+  type ImmutableBitSetCC[X] = ({ type L[_] = i.BitSet })#L[X]
+  type MutableBitSetCC[X] = ({ type L[_] = m.BitSet })#L[X]
 }
