+++ src/main/scala-2.11_2.12/scala/collection/compat/immutable/ArraySeq.scala
@@ -1,4 +1,4 @@
-package scala.collection.compat.imm
+package scala.collection.compat.immutable
 
 import java.util.Arrays
