+++ src/main/scala-2.13/scala/collection/compat/immutable/package.scala
@@ -1,6 +1,6 @@
 package scala.collection.compat
 
-package object imm {
+package object immutable {
   type ArraySeq[+T] = scala.collection.immutable.ArraySeq[T]
   val ArraySeq      = scala.collection.immutable.ArraySeq
 }
