+++ src/main/scala-2.13/scala/collection/compat/package.scala
@@ -2,10 +2,11 @@ package scala.collection
 
 package object compat {
   type Factory[-A, +C] = scala.collection.Factory[A, C]
-  val Factory          = scala.collection.Factory
+  val  Factory         = scala.collection.Factory
 
   type BuildFrom[-From, -A, +C] = scala.collection.BuildFrom[From, A, C]
-  val BuildFrom                 = scala.collection.BuildFrom
+  val  BuildFrom                = scala.collection.BuildFrom
 
   type IterableOnce[+X] = scala.collection.IterableOnce[X]
+  val  IterableOnce     = scala.collection.IterableOnce
 }
